[PUBDEV-3616] move water.rapids.Val => water.rapids.vals.Val
