+++ h2o-core/src/main/java/water/rapids/ast/AstFunction.java
@@ -2,7 +2,7 @@ package water.rapids.ast;
 
 import water.H2O;
 import water.rapids.Env;
-import water.rapids.Val;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValFun;
 import water.util.StringUtils;
