+++ h2o-core/src/main/java/water/rapids/ast/params/AstStrList.java
@@ -1,8 +1,7 @@
 package water.rapids.ast.params;
 
 import water.rapids.Env;
-import water.rapids.Rapids;
-import water.rapids.Val;
+import water.rapids.vals.Val;
 import water.rapids.ast.AstParameter;
 
 import java.util.ArrayList;
