+++ h2o-core/src/main/java/water/rapids/ast/prims/math/AstUniOp.java
@@ -5,6 +5,7 @@ import water.fvec.*;
 import water.rapids.*;
 import water.rapids.ast.Ast;
 import water.rapids.ast.AstFunction;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValFrame;
 import water.rapids.vals.ValNum;
 import water.rapids.vals.ValRow;
