+++ h2o-core/src/main/java/water/rapids/ast/prims/mungers/AstApply.java
@@ -6,6 +6,7 @@ import water.MRTask;
 import water.rapids.*;
 import water.rapids.ast.*;
 import water.rapids.ast.Ast;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValFrame;
 
 /**
