+++ h2o-core/src/main/java/water/rapids/ast/prims/operators/AstIfElse.java
@@ -9,6 +9,7 @@ import water.fvec.Vec;
 import water.rapids.*;
 import water.rapids.ast.AstFunction;
 import water.rapids.ast.Ast;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValFrame;
 import water.rapids.vals.ValNum;
 import water.rapids.vals.ValRow;
