+++ h2o-core/src/main/java/water/rapids/ast/prims/reducers/AstRollupOp.java
@@ -3,7 +3,7 @@ package water.rapids.ast.prims.reducers;
 import water.fvec.Frame;
 import water.fvec.Vec;
 import water.rapids.Env;
-import water.rapids.Val;
+import water.rapids.vals.Val;
 import water.rapids.ast.Ast;
 import water.rapids.vals.ValNum;
 import water.rapids.vals.ValRow;
