+++ h2o-core/src/main/java/water/rapids/ast/prims/repeaters/AstRepLen.java
@@ -5,7 +5,7 @@ import water.fvec.Chunk;
 import water.fvec.Frame;
 import water.fvec.Vec;
 import water.rapids.Env;
-import water.rapids.Val;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValFrame;
 import water.rapids.ast.AstFunction;
 import water.rapids.ast.Ast;
