+++ h2o-core/src/main/java/water/rapids/vals/Val.java
@@ -1,4 +1,4 @@
-package water.rapids;
+package water.rapids.vals;
 
 import water.fvec.Frame;
 import water.Iced;
