+++ h2o-core/src/main/java/water/rapids/vals/ValFrame.java
@@ -1,7 +1,6 @@
 package water.rapids.vals;
 
 import water.fvec.Frame;
-import water.rapids.Val;
 
 /**
  * Value that represents an H2O dataframe ({@link Frame}).
