+++ h2o-core/src/main/java/water/rapids/vals/ValNum.java
@@ -1,7 +1,5 @@
 package water.rapids.vals;
 
-import water.rapids.Val;
-
 /**
  * Numeric value. We do not distinguish between integers and floating point numbers.
  */
