+++ h2o-core/src/test/java/water/rapids/RBindTest.java
@@ -7,6 +7,7 @@ import water.Key;
 import water.Keyed;
 import water.TestUtil;
 import water.fvec.Frame;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValFrame;
 
 public class RBindTest extends TestUtil {
