+++ h2o-core/src/test/java/water/rapids/RapidsTest.java
@@ -12,6 +12,7 @@ import water.parser.ParseSetup;
 import water.rapids.ast.Ast;
 import water.rapids.ast.params.AstNumList;
 import water.rapids.ast.params.AstStr;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValFrame;
 import water.util.ArrayUtils;
 import water.util.Log;
