+++ h2o-core/src/test/java/water/rapids/ast/prims/mungers/AstGetrowTest.java
@@ -6,7 +6,7 @@ import water.TestUtil;
 import water.fvec.Frame;
 import water.fvec.Vec;
 import water.rapids.Rapids;
-import water.rapids.Val;
+import water.rapids.vals.Val;
 import water.rapids.vals.ValRow;
 import water.util.ArrayUtils;
